+++ tests/unit/helpers/IbanTest.php
@@ -10,7 +10,7 @@ final class IbanTest extends TestCase
 {
     public function testValidIbans(): void
     {
-        self::assertTrue(Iban::isValid('HR1210010051863000160'));
+        self::assertTrue(Iban::isValid('HR3799999990000000001'));
         self::assertTrue(Iban::isValid('hr12 1001 0051 8630 0016 0'));
         self::assertTrue(Iban::isValid('DE89370400440532013000'));
         self::assertTrue(Iban::isValid('SI56020170014356205'));